cleanup client and move Peer to separate file
